+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Support embassy-executor v0.10.x via feature `embassy010`
+
 ## [0.7.0]
 
 ### Added